Upgraded to Ruby 2.5.1 and Rails 5.2.
Upgraded Ruby to 2.5.1.
Upgraded Rails to 5.2 and ran update script which updated a bunch of files. It tends to overwrite files so I manually edited each file so that nothing was lost.
A few files outside of those directly in the upgrades were touched to get the one test to pass. Also, the Gemfile is updated and organized, adding in the various gems needed to support testing, as well as the ones mentioned in email.