Removed fixtures.
I originally setup the tests to use AnyFixture and share factories across specs. This proved to be very problematic. RSpec isn't really designed to setup data and use it repeatedly so I removed the fixtures and used factories in their place. This means the tests will hit the database a bit more and will be a bit slower but really it's a negligible difference.